Expose usable interface
While the test passed before, it was impossible to use the API since
Websocket itself was not exported. By following a similar pattern to
websocket_lwt and websocket_async, this commit allows Frame to be
exposed correctly, and thus enables upgrade_connection to be used in
external applications.

This commit also updates the test accordingly, which now should not rely
on any internal functionality.